feat: adds editor config, replaces some nugets, adds xru as linked file
