+++ xUnitRevitUtils2019/xUnitRevitUtils2019.csproj
@@ -7,10 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\xUnitRevitUtils2021\xru.cs" Link="xru.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="ModPlus.Revit.API.2019" Version="4.0.0" />
     <PackageReference Include="xunit" Version="2.4.2" />
   </ItemGroup>
