+++ xUnitRevitUtils2022/xUnitRevitUtils2022.csproj
@@ -6,8 +6,9 @@
     <PackageId>xUnitRevitUtils.2022</PackageId>
   </PropertyGroup>
 
+
   <ItemGroup>
-    <PackageReference Include="ModPlus.Revit.API.2022" Version="4.0.0" />
+    <PackageReference Include="Speckle.Revit.API" Version="2022.0.2.1" />
     <PackageReference Include="xunit" Version="2.4.2" />
   </ItemGroup>
